# Conflicts:
#	TBEA/BusinessPerformance/Documentation/DB sql/1.26/权限.sql

+++ TBEA/BusinessPerformance/Web/testWebProject/src/main/java/com/tbea/ic/operation/model/dao/authority/ExtendAuthorityDaoImpl.java
@@ -39,7 +39,7 @@ public class ExtendAuthorityDaoImpl  implements ExtendAuthorityDao{
 		Query q = this.entityManager.createNativeQuery("select count(*) from system_extend_auth_v where account_id = :id and authId = :auth");
 		q.setParameter("id", account.getId());
 		q.setParameter("auth", auth);
-		return ((Long)q.getResultList().get(0)).intValue();
+		return (Integer)q.getResultList().get(0);
 	}
 
 	@Override
